cleaned up index file
